pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "private-domain-ai-brain"
version = "0.1.0"
description = "私域运营专家 AI 智脑助手 - 基于 LangGraph Supervisor 模式的多智能体系统"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "deepagents>=0.4.11",

    # API Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",

    # Database
    "psycopg[binary,pool]>=3.2.0",
    "asyncpg>=0.30.0",

    # Vector DB
    "pymilvus>=2.4.0",

    # File Processing
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "python-docx>=1.1.0",
    "pypdf>=5.0.0",

    # Data Analysis
    "matplotlib>=3.9.0",
    "plotly>=5.24.0",
    "kaleido>=0.2.1",
    "numpy>=2.1.0",
    "scipy>=1.14.0",

    # Embeddings / Reranker
    "sentence-transformers>=3.3.0",
    "FlagEmbedding>=1.3.0",

    # Utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "tiktoken>=0.8.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
